+++ test/examples/chat/server.ts
@@ -8,7 +8,6 @@ import { chatRouter } from './routes/chat';
 import { ChatMessage } from './models/chat-message';
 import { User } from './models/user';
 
-const ROOT_PATH = require('app-root-path').toString();
 pointy.userType = User;
 
 // Setup
@@ -27,7 +26,7 @@ pointy.before = async (app) => {
 	// Database
 	await pointy.db
 		.setEntities([ User, ChatMessage ])
-		.connect(ROOT_PATH)
+		.connect()
 		.catch((error) => pointy.error(error));
 };