+++ test/spec/chat/server.spec.ts
@@ -4,7 +4,6 @@ import { pointy } from '../../../src/';
 import { forkServer } from '../../../src/utils/fork-server';
 import { User } from '../../examples/chat/models/user';
 import { ChatMessage } from '../../examples/chat/models/chat-message';
-const ROOT_PATH = require('app-root-path').toString();
 
 const http = pointy.http;
 
@@ -18,7 +17,7 @@ beforeAll(async () => {
 	// Database
 	await pointy.db
 		.setEntities([ ChatMessage, User ])
-		.connect(ROOT_PATH)
+		.connect()
 		.catch((error) =>
 			fail('Cannot start database' + JSON.stringify(error))
 		);