+++ test/spec/terms/server.spec.ts
@@ -4,7 +4,6 @@ import { pointy } from '../../../src/';
 import { forkServer } from '../../../src/utils/fork-server';
 import { User } from '../../examples/terms/models/user';
 import { Term } from '../../examples/terms/models/term';
-const ROOT_PATH = require('app-root-path').toString();
 
 const http = pointy.http;
 
@@ -16,7 +15,7 @@ beforeAll(async () => {
 	// Database
 	await pointy.db
 		.setEntities([ Term, User ])
-		.connect(ROOT_PATH)
+		.connect()
 		.catch((error) =>
 			fail('Cannot start database' + JSON.stringify(error))
 		);