+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Breaking Changes (See migration.md)
 - Upgrade to node v16 & npm v8
 - Upgrade to class validator
+- [Issue #202] Switch to .env from local.config.json
 
 ### Additions